chore: narrow PR to LTT docs, tests, and scheduling code only
Remove scripts and unrelated config churn; keep decrement/reset on
slot_bookings, availability capacity field, LTT pytest suites, and
AsciiDoc write-ups for #505 and #506.
