+++ backend/app/routes/main.py
@@ -3,14 +3,21 @@ from datetime import date
 from flask import Blueprint, jsonify, request
 
 from app.middleware.auth_middleware import requires_auth
-from app.services.slot_capacity import slot_time_strings_for_doctor_day
+from app.repositories.slot_bookings import slot_time_strings_for_doctor_day
 
+# A Blueprint groups related routes together.
+# 'main' is the name used internally by Flask to identify this blueprint.
+# To add more route files: create a new file (e.g. appointments.py), define a blueprint,
+# then import and register it in routes/__init__.py.
 main_bp = Blueprint('main', __name__)
 
 
 @main_bp.route('/health', methods=['GET'])
 def health_check():
-    """Health check endpoint — confirms the API is running."""
+    """Health check endpoint — confirms the API is running.
+    Useful for deployment checks and debugging connectivity.
+    Access at: GET /api/health
+    """
     return jsonify({'status': 'ok'})
 
 
@@ -23,7 +30,7 @@ def get_doctor_available_slots(doctor_id: str):
         date: YYYY-MM-DD (required)
 
     Returns JSON: {"slots": ["09:00", "09:30", ...]} — slots inside working hours,
-    respecting breaks, duration, and per-slot capacity.
+    respecting breaks and appointment duration. Empty list with 200 OK if none.
     """
     date_str = request.args.get('date')
     if not date_str:
